requirements.txt
amqp==2.5.0
autopep8==1.4.4
backports.csv==1.0.7
beautifulsoup4==4.8.1
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
et-xmlfile==1.0.1
future==0.17.1
grpcio==1.22.0
grpcio-tools==1.22.0
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
joblib==0.13.2
kombu==4.6.3
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.4
odfpy==1.4.0
openpyxl==2.6.2
protobuf==3.8.0
pycodestyle==2.5.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
